refactor(cli): move makeTmpDirTracker into shell
Routing every teardown through the tracker made
src/domains/flows/ensureDeps.test.ts import it from
src/domains/runtimeEnv, and a domain must not import a sibling domain.
The tracker only wraps mkdtemp and rm, so it belongs in shell. No
override in .oxlintrc.json covers runtimeEnv, so lint did not catch it.
