+++ src/domains/runtimeEnv/outerHop.test.ts
@@ -1,3 +1,4 @@
+import { makeTmpDirTracker } from "~/shell/tmpDir.testUtils.js";
 import { afterEach, describe, expect, it } from "bun:test";
 import { writeFile } from "node:fs/promises";
 import { join } from "node:path";
@@ -8,7 +9,6 @@ import { populateOuterHop } from "./outerHop.js";
 import {
   makeInstallMock,
   makeProjectTree,
-  makeTmpDirTracker,
 } from "./runDirFixtures.testUtils.js";
 import { scaffoldManagedRuntime } from "./scaffoldManagedRuntime.testUtils.js";
 import { expectLinkTarget } from "./symlinkDir.testUtils.js";
