+++ src/domains/runtimeEnv/prepareRunDir.outerHop.test.ts
@@ -1,12 +1,10 @@
+import { makeTmpDirTracker } from "~/shell/tmpDir.testUtils.js";
 import { afterEach, describe, expect, it } from "bun:test";
 import { writeFile } from "node:fs/promises";
 import { join } from "node:path";
 
 import { prepareRunDir } from "./prepareRunDir.js";
-import {
-  makeProjectTree,
-  makeTmpDirTracker,
-} from "./runDirFixtures.testUtils.js";
+import { makeProjectTree } from "./runDirFixtures.testUtils.js";
 import { scaffoldManagedRuntime } from "./scaffoldManagedRuntime.testUtils.js";
 import { expectLinkTarget } from "./symlinkDir.testUtils.js";
